댓글 delvelop(고정/관리자 삭제기능)

+++ sec_mini/src/main/java/controller/CommentController.java
@@ -4,6 +4,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -16,6 +17,7 @@ import dao.CommentLikesDao;
 import util.MyCommon;
 import util.Paging;
 import vo.CommentVo;
+import vo.UserVo;
 
 @Controller
 @RequestMapping("/comment/")
@@ -65,11 +67,43 @@ public class CommentController {
 
     @RequestMapping(value="delete.do", produces="application/json; charset=utf-8;")
     @ResponseBody
-    public String delete(int cmt_idx) {
-        commentLikesDao.deleteByComment(cmt_idx); // 댓글의 모든 좋아요 삭제
-        int res = comment_dao.delete(cmt_idx);
+    public String delete(int cmt_idx, HttpSession session) {
+        UserVo user = (UserVo) session.getAttribute("user");  // "loggedInUser" -> "user"로 변경
+        
+        System.out.println("Logged in user: " + user); // 로그 추가
+
         JSONObject json = new JSONObject();
-        json.put("result", res == 1); 
+        
+        if (user == null) {
+            json.put("result", "failure");
+            json.put("message", "로그인이 필요합니다.");
+            return json.toString();
+        }
+
+        CommentVo comment = comment_dao.selectByIdx(cmt_idx);
+
+        System.out.println("Comment: " + comment); // 로그 추가
+        
+        if (comment == null) {
+            json.put("result", "failure");
+            json.put("message", "댓글을 찾을 수 없습니다.");
+            return json.toString();
+        }
+
+        if (user.getUserNo() == comment.getUserNo()) {
+            // 자신의 댓글인 경우 실제로 삭제
+            commentLikesDao.deleteByComment(cmt_idx); // 댓글의 모든 좋아요 삭제
+            int res = comment_dao.delete(cmt_idx);
+            json.put("result", res == 1);
+        } else if ("관리자".equals(user.getNickName())) {
+            // 관리자인 경우 is_deleted 플래그 업데이트
+            int res = comment_dao.markCommentAsDeleted(cmt_idx);
+            json.put("result", res == 1);
+        } else {
+            json.put("result", "failure");
+            json.put("message", "삭제 권한이 없습니다.");
+        }
+
         return json.toString();
     }
